+++ src/files-sidebar-hooks.js
@@ -17,35 +17,23 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import Vue from 'vue';
 import { appName } from './config.js';
-import { attachDialogHandlers } from './toolkit/util/dialogs.js';
+import { translate as t } from '@nextcloud/l10n';
 import { getInitialState } from './toolkit/services/InitialStateService.js';
-import { generateUrl } from '@nextcloud/router';
-import FilesTab from './views/FilesTab.vue';
-import { Tooltip } from '@nextcloud/vue';
 
 // eslint-disable-next-line
 import logoSvg from '../img/app.svg?raw';
 
-require('files-archive.scss');
-require('dialogs.scss');
-
-Vue.directive('tooltip', Tooltip);
+// require('files-archive.scss'); // ? still needed ?
 
 // eslint-disable-next-line
 require('./webpack-setup.js');
 
-Vue.mixin({ data() { return { appName }; }, methods: { t, n, generateUrl } });
-
-const View = Vue.extend(FilesTab);
 let TabInstance = null;
 
 const initialState = getInitialState();
 
 window.addEventListener('DOMContentLoaded', () => {
-
-  attachDialogHandlers();
 
   /**
    * Register a new tab in the sidebar
@@ -61,6 +49,11 @@ window.addEventListener('DOMContentLoaded', () => {
       },
 
       async mount(el, fileInfo, context) {
+        console.info('FILES ARCHIVE SIDEBAR LOAD');
+        const FilesTabAsset = (await import('./views/FilesTab.vue'));
+        const Vue = FilesTabAsset.Vue;
+        const FilesTab = FilesTabAsset.default;
+        const View = Vue.extend(FilesTab);
 
         if (TabInstance) {
           TabInstance.$destroy();
@@ -73,14 +66,15 @@ window.addEventListener('DOMContentLoaded', () => {
 
         // Only mount after we have all the info we need
         await TabInstance.update(fileInfo);
-
         TabInstance.$mount(el);
       },
       update(fileInfo) {
         TabInstance.update(fileInfo);
       },
       destroy() {
-        TabInstance.$destroy();
+        if (TabInstance !== null) {
+          TabInstance.$destroy();
+        }
         TabInstance = null;
       },
     }));