+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.2.0-rc1] - unreleased
+## [1.2.0] - 2024-03-17
 
 ### Added
 
@@ -13,8 +13,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Reduce the size of the JS assets needed to hook in to the files-app sidebar
+
+- Use latest @nextcloud/vue toolkit.
+
 - Drop support for Nextcloud <= v27. The differences in the files-API
   are just too big.
+
+- Translations
 
 ### Fixed